기능 개선 : 수의 범위를 int형에서 long형으로 확장

+++ src/main/java/com/caculator/StringUtils.java
@@ -4,7 +4,7 @@ public class StringUtils {
 
     public static boolean isNumber(String s) {
         try {
-            Integer.parseInt(s);
+            Long.parseLong(s);
             return true;
         } catch (NumberFormatException e) {
             return false;
